requirements.txt
Babel==2.9.0
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.0.0
MarkupSafe==2.1.1
python-dateutil==2.6.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.34
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.7.0